requirements.txt
pandas
sqlalchemy
pyyaml
psycopg2-binary
python-dotenv
pytest
black
isort
flake8
